fix: re-organize test utility functions
