+++ tests/postgresOpsTests.js
@@ -18,8 +18,8 @@ const jqUnit = require("node-jqunit");
 const postgresOps = require("../src/dbOps/postgresOps.js");
 
 require("../src/shared/driverUtils.js");
-require("./shared/utilsCommon.js");
-require("./shared/utilsDb.js");
+require("./shared/testUtils.js");
+require("./shared/testDbUtils.js");
 
 // Tables structures and test data records
 require("./data/testTableModels.js");
