+++ tests/postgresRequestTests.js
@@ -15,8 +15,7 @@ const jqUnit = require("node-jqunit");
 const postgresOps = require("../src/dbOps/postgresOps.js");
 
 require("../src/shared/driverUtils.js");
-require("./shared/utilsCommon.js");
-require("./shared/utilsDb.js");
+require("./shared/testUtils.js");
 
 jqUnit.module("PostgresDB request unit tests.");
