+++ tests/tableModelDefTests.js
@@ -15,8 +15,8 @@ const jqUnit = require("node-jqunit");
 const postgresOps = require("../src/dbOps/postgresOps.js");
 
 require("../src/shared/driverUtils.js");
-require("./shared/utilsCommon.js");
-require("./shared/utilsDb.js");
+require("./shared/testUtils.js");
+require("./shared/testDbUtils.js");
 
 jqUnit.module("PostgresDB table definitions unit tests.");
